fix: Issues identified in RTE implementation
- Updated promote to be tollerant of emtpy results
- Added missing logic from prev implementation. These are applying the
applyOrCut and applyAndCut functions to children and applying promote
after reduceToElegance applyed to Or nodes.
- One failing test case from python impelmentation. To address this the
applyReduceToEleganceToOr and applyReduceToEleganceToAnd were refactored
so that they execute keeping the reset logic under consideration; the
zeroConstraintSubsume function variation to apply for a given POA is
refactored to work with OR nodes too and called in the OR case of the
RTE to remove an empty OR node if exists; fixed incorrect state passing
to the applyOrCut and applyAndCut which resulted in unusual results.
- Fixed the issue of removing OR nodes with no children but have
literals